+++ VendingMachine.Coin/CoinService.cs
@@ -2,16 +2,16 @@
 {
     public class CoinService : ICoinService
     {
-        public Coin GetCoin(decimal weight, decimal diameter, decimal thickness)
+        public Coin GetCoin(double weight, double diameter, double thickness)
         {
             // This should have acceptable legal tolerances in here
-            if (weight == 2 && diameter == 21.12m && thickness == 1.95m) return Coin.Nickel;
+            if (weight == 2 && diameter == 21.12 && thickness == 1.95) return Coin.Nickel;
             
-            if (weight == 2.5m && diameter == 19.05m && thickness == 1.52m) return Coin.Penny;
+            if (weight == 2.5 && diameter == 19.05 && thickness == 1.52) return Coin.Penny;
             
-            if (weight == 6.25m && diameter == 24.26m && thickness == 1.75m) return Coin.Quarter;
+            if (weight == 6.25 && diameter == 24.26 && thickness == 1.75) return Coin.Quarter;
             
-            if (weight == 2.268m && diameter == 17.91m && thickness == 1.35m) return Coin.Dime;
+            if (weight == 2.268 && diameter == 17.91 && thickness == 1.35) return Coin.Dime;
 
             return Coin.Invalid;
         }
